Android detail edit

+++ DataDamoa/app/src/main/java/com/example/datadamoa/Work.java
@@ -3,5 +3,5 @@ package com.example.datadamoa;
 public class Work {
     public int idx;
     public String title;
-    public double price;
+    public int price;
 }
